requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
jinja2
python-multipart
pydantic
python-dotenv
itsdangerous
google-generativeai
apscheduler
fastapi-mail
